requirements.txt
# XMAS 1Week Event System - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
cryptography==41.0.7

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation (Python 3.13 compatible - use pre-built wheels)
pydantic>=2.9.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Timezone
pytz==2023.3
tzdata>=2024.1

# HTTP Client (optional, for external API calls)
httpx==0.26.0

# Redis (optional, for caching)
redis==5.0.1

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # for TestClient

# Logging & Monitoring (optional)
# sentry-sdk==1.39.2
